Added user-definable reload actions to the devel-handler
Also added a demo file demonstrating a compelling use case: The user
clicks on the browser the first time the app loads. Each successive time
the server is reloaded, the browser is automatically refreshed (sent the
F5 key). This example requires xdotool (linux only?) to collect the
window pid and send the F5 keypress.