serverHeader does not add Server: if already exists.
serverHeader was added by me when I uses Warp as a server only.
This function adds the Server: field if not exists and appends
values if exists.

After start using it as a proxy server, it appeared that this
feature is inconvenient. A proxy server should not add the
Server: field according to specification. (It should add the Via:
field.)

So, let's not add the Server: field if exists. This works in many
cases.