RequestBody is now a Source
This may not be a good move, but here's the rationale. We have a few
reasons to avoid buffering:

* It's a performance overhead
* It's less well understood

In this case, none of our code is actually utilizing the buffered aspect
of the source. Warp comes the closest, but in fact only uses it
internally. After it's done with the buffered actions, it applies
isolate, which creates an unbuffered source again.

Overall, I'm beginning to think that the place for buffering is only
internal, and all external interfaces should be unbuffered (there's a
similar commit to http-conduit).