wai-eventsource:  Do not a blaze-builder flush on eventToBuilder.
Adding a blaze-builder flush there has the disadvantage of
creating an HTTP chunk for each event on the response.  OTOH, it
has no advantage: a wai Flush is added after each wai Chunk by
all functions from the EventSource module, so the blaze-builder
flush is redundant.

Note that this commit does not change the behaviour of the
EventSource module at all.  It only changes the behaviour of any
3rd party code using EventStream (which I assume is only mine's
for now :]).