Add allow-sendfilefd flag for Debian GNU/kFreeBSD
On Debian GNU/kFreeBSD (FreeBSD kernel plus glibc), the BSD flavour of
sendfile is not usable, but Cabal treats kfreebsdgnu as os(freebsd) so I
had to add a flag to simple-sendfile.cabal to make it possible to select
the fallback implementation.  See
https://github.com/kazu-yamamoto/simple-sendfile/pull/13 for more
details.

warp.cabal reflects the availability of interfaces from simple-sendfile,
and so I believe it needs a matching flag to make it possible to
forcibly disable the use of -DSENDFILEFD.