Handle a timeout bug during request body
Here's the problem situation.  As soon as we are done with processing
the request headers, the timeout handler is paused. The purpose is to
prevent timeouts during application code. The problematic case this
commit solves is a when the request body has at least one chunk >= 2048.
In this case, this chunk will tickle the timeout handler, which will
start the clock again. If the application then delays longer than the
timeout period, it will be terminated.

Instead, we resume the timeout handler as soon as we start reading the
request body, and then pause it as soon as we exit the request body
processing code.

Note that if the Sink provided by the application takes a long time, we
could theoretically still have a problem. This should be looked into in
more detail at some point.