auto-update: only ever fork one worker thread
Goal: fork a single worker thread immediately, and have it sleep when
it's not needed.

Current code has a downside that after delay, the next time the value is
demanded there will be a lag before the value is available. That can be
fixed with an IORef, but I need to check if that makes performance
worse.