Only conditionally produce HTTP 100 Continue
Only produce HTTP 100 Continue if the application asks for the request
body. This allows the application to respond with a 4xx or 5xx status,
having only looked at the request headers, and never receiving the body.
An example use-case for this would be responding with HTTP 403 or 413,
without having to waste resources sending a large request body.