Have stream bodies just return trailers directly
The 'TrailerFunc' approach was needlessly cumbersome since its power can
be recovered with a combination of 'try' and '>>='.  This way it's more
clear what will happen to exceptions in the stream body: if they're
uncaught, they'll escape into Warp and we'll reset the stream; if they
are caught and the stream body returns trailers, Warp can't know about
the exception, so it will necessarily send them and not reset the stream.