Improvements to ping's javascript script.
- Avoid using the browser's cache for the response. IE returns the
response from cache and does not send the ping request to the server
causing the server to shut down.

- Make the javascript use the asynchronous ajax request instead of
synchronous. This fixes the warning 'Synchronous XMLHttpRequest on the
main thread is deprecated because of its detrimental effects to the end
user's experience. For more help, check http://xhr.spec.whatwg.org/.'
displayed in the Chrome browser.