don't launch if server fails; kill server on exit #537
Make the server thread and launch/monitor thread more aware of each
other, using async. Now, we won't launch a browser until the server is
ready to serve pages, and we won't launch it at all if the server fails
to start. Also we make sure to terminate the server thread whenever the
main thread terminates, eg from ctrl-c in GHCI.