warp: add HTTP2 CONNECT support
As per section 8.3 of HTTP2 spec, ":scheme" and ":path" should be
omitted and ":authority" is mandatory for HTTP2 CONNECT request.

Treat CONNECT request as a special case for headers check. And also
for CONNECT request, use ":authority" as unparsed path, so that
CONNECT method has consistent behavior compare to HTTP 1.0.

Closes #561