Add output formatter with details and headers
This adds a custom log output formatter which includes response headers.
The motivation for this is that I'd like to be able to add arbitrary
and request-specific data from my [Yesod] application to the log output,
and it seems the best way to do this is to add that data to response
headers.