Added ssUnhandledRequestHandler
wai-app-static used to summarily reject any request that
was not a GET or HEAD. This made it very difficult to use
it as a "middleware" of sorts in front of anything that was
expecting POST/PUT/DELETE requests, eg. a Scotty application.

ssUnhandledRequestHandler gives an optional  hook (similar to
ss404Handler) to allow wai-app-static to pass on non GET/HEAD
requests to another Wai application.