Document the Middleware type alias
I tried to come at this from the angle of

1) A user of Wai trying to understand what `Middleware` is
   and how to use it, especially what happens if you chain
   multiple middlewares and in what order their are called.

2) A developer trying to implement new middleware, how to
   modify requests and responses, and how to pass on data.

by giving examples of how the type signatures and
simple usage/implementations look.

I think `Wai` is a super fundamental module for the Haskell
ecosystem, so we should aim for the documentation to be top
notch (think: rust stdlib).

I think the rest of the module needs a simple overhaul,
but I was especially annoyed by having to reconstruct Middleware
from first principles every time I built a server (“which order
does it chain again?”).