Install settings shutdown handler for TLS servers
Any shutdown handler passing in the settings to 'runTLS' or
'runTLSSocket' was not being installed in the same way that
'runSettingsSocket' and friends do for non-TLS servers.

This updates 'runTLSSocket' to make it install the shutdown handler from
the settings so that warp-tls behaves the same way as warp in this
regard.