Make reaper workloads modifiable
It is convenient to be able to modify reaper workloads outside of the
`reaperAction` in scenarios where we don't want to wait for `reaperDelay` for
the jobs to be updated. This PR adds an atomic `reaperModify` function to the
reaper API which enables this.

Since this changes the visible `Reaper` constructor, I did a major version bump
(`auto-update` 0.1.6 -> 0.2.0) and added a changelog entry. I also bumped the
`warp` dependency on `auto-update`.