Eagerly remove handles from the manager on `cancel`
`cancel` used to mark the given `Handle` as `Canceled`, which relied on the
reaper to run and clean up the `Canceled` handles. This is not good for long
timeouts when many streams are opened, as a lot of `Canceled` handles build up
in the reaper's workload. This commit uses the reaper's new `reaperModify` API
to immediately remove the given `Handle` in `cancel`. This obsoletes the
`Canceled` constructor of `State`, so we remove that as well.

We update the time-manager dependency of warp to include the new changes as
well.

The related auto-update changes are in PR #985.