prevent calling "connClose" twice.
TimeManager is leaking because of "BlockedOnMVar".
Both TimeManger and Reaper do not use MVar at all.
TLS "bye" is the only registered function which uses MVar.
When TimeManager kills Warp, "bye" is called twice.
One is via "bracket", the other is by TimeManager.
This ensures that "bye" is called only once.